requirements.txt
asgiref==3.2.3
astroid==2.3.3
certifi==2020.4.5.1
chardet==3.0.4
confusable-homoglyphs==3.2.0
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.42.0
django-crispy-forms==1.9.2
django-registration==3.1
django-rest-auth==0.9.5
django-webpack-loader==0.7.0
djangorestframework==3.11.0
# get==2019.4.13
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.1.1
# post==2019.4.13
# public==2019.4.13
pylint==2.4.4
python3-openid==3.1.0
# pytz==2019.3
# query-string==2019.4.13
# request==2019.4.13
requests==2.23.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.13.0
sqlparse==0.3.0
stripe==2.45.0
typed-ast==1.4.0
urllib3==1.25.8
virtualenv==16.7.9
wrapt==1.11.2
